moved services into seperate files, and put down match-case statements when handling Result types
