+++ launcher/components/install.py
@@ -20,20 +20,21 @@ def _validated_apworld(apworld_src: str = "") -> tuple[Path, str, str] | None:
     """
 
     if not apworld_src:
-        apworld_src = open_filename("Select APWorld file to install", (("APWorld", (".apworld",)),))
-        if not apworld_src:
+        selected_apworld = open_filename("Select APWorld file to install", (("APWorld", (".apworld",)),))
+        if not selected_apworld:
             return None
+        apworld_src = selected_apworld
 
     result = dispatch(ValidateInstall(apworld_path=apworld_src))
-    if isinstance(result, Err):
-        raise Exception(result.error.message)
-    assert isinstance(result, Ok)
-    if not result.value.valid:
-        raise Exception(result.value.error or f"APWorld is not installable: {apworld_src}")
-    if not result.value.module_name or not result.value.apworld_name:
-        raise Exception("Validated APWorld is missing module metadata.")
-
-    return Path(apworld_src), result.value.module_name, result.value.apworld_name
+    match result:
+        case Err(error=error):
+            raise Exception(error.message)
+        case Ok(value=value):
+            if not value.valid:
+                raise Exception(value.error or f"APWorld is not installable: {apworld_src}")
+            if not value.module_name or not value.apworld_name:
+                raise Exception("Validated APWorld is missing module metadata.")
+            return Path(apworld_src), value.module_name, value.apworld_name
 
 
 def _install_apworld(apworld_src: str = "") -> tuple[Path, Path] | None:
@@ -46,10 +47,11 @@ def _install_apworld(apworld_src: str = "") -> tuple[Path, Path] | None:
         apworld_src = str(validated[0])
 
     result = dispatch(InstallApworld(apworld_path=apworld_src))
-    if isinstance(result, Err):
-        raise Exception(result.error.message)
-    assert isinstance(result, Ok)
-    return Path(result.value.source_path), Path(result.value.target_path)
+    match result:
+        case Err(error=error):
+            raise Exception(error.message)
+        case Ok(value=value):
+            return Path(value.source_path), Path(value.target_path)
 
 
 def install_apworld(apworld_path: str = "") -> None:
